xtracted github release gulp task in own file used node-fetch instead of https
