+++ gulpfile.ts
@@ -5,9 +5,7 @@
 */
 
 import gulp from 'gulp'
-import https from 'https'
-import fs from 'fs/promises'
-import yaml from 'yaml'
+import {githubReleases} from "./github-releases";
 
 gulp.task("copy:font:lato", async () => {
   gulp.src([
@@ -29,79 +27,8 @@ gulp.task("copy:font:fork-awesome", async () => {
   gulp.src('./node_modules/fork-awesome/fonts/*').pipe(gulp.dest('./static/fonts/forkawesome'))
 });
 
-interface GitHubReleaseEntry {
-  id: string
-  name: string
-  tag_name: string
-  html_url: string
-  tarball_url: string
-  body: string
-  assets: {
-    name: string
-    size: number,
-    browser_download_url: string
-  }[]
-  created_at: string
-  prerelease: boolean
-  draft: boolean
-}
-
 gulp.task("github-releases", async () => {
-
-  const options = {
-    headers: {
-      'User-Agent': 'Mozilla/5.0'
-    }
-  }
-
-  https.get(' https://api.github.com/repos/hedgedoc/hedgedoc/releases', options, (response: { on: (arg0: string, arg1: (chunk: string) => void) => void; }) => {
-    let data = '';
-
-    // A chunk of data has been recieved.
-    response.on('data', (chunk) => {
-      data += chunk;
-    });
-
-    // The whole response has been received. Print out the result.
-    response.on('end', () => {
-      let apiResult: GitHubReleaseEntry[] = JSON.parse(data)
-        .filter((entry: GitHubReleaseEntry) => !entry.prerelease && !entry.draft)
-
-      // bring releases in order
-      //apiResult = apiResult.sort((a: GitHubReleaseEntry, b: GitHubReleaseEntry) => semver.gte(a.tag_name, b.tag_name) ? -1 : 1);
-
-      apiResult.forEach(entry => {
-        const frontmatterJson = {
-          Title: entry.name,
-          date: entry.created_at,
-          tarball: entry.tarball_url,
-          assets: entry.assets.map(asset => {
-            return {
-              name: asset.name,
-              url: asset.browser_download_url,
-              size: asset.size
-            }
-          }),
-          githubLink: entry.html_url
-        };
-
-        const content = `---
-${yaml.stringify(frontmatterJson)}
----
-${entry.body}`;
-
-        const filename = `content/english/releases/${entry.tag_name}.md`;
-
-        fs.access(filename).catch(() => {
-          console.log("write file ", filename);
-          fs.writeFile(filename, content);
-        });
-      });
-    });
-
-  }).on("error", (error: Error) => {
-    console.log("Error: " + error.message);
-  });
+  await githubReleases();
 });
 
 gulp.task("default", gulp.parallel("copy:font:lato", "copy:font:kumbh-sans", "copy:font:fork-awesome", "github-releases"))
