+++ package.json
@@ -11,12 +11,14 @@
     "bootstrap": "4.5.3",
     "fork-awesome": "1.1.7",
     "gulp": "4.0.2",
+    "node-fetch": "^2.6.1",
     "ts-node": "9.1.1",
     "typescript": "4.1.3",
     "yaml": "^1.10.0"
   },
   "devDependencies": {
-    "@lhci/cli": "0.6.1"
+    "@lhci/cli": "0.6.1",
+    "@types/node-fetch": "^2.5.7"
   },
   "scripts": {
     "prepare": "gulp"
